+++ src/connector/api/controller/explain_controller.rs
@@ -29,7 +29,8 @@ impl<'a> ExplainController<'a> {
         dump_symbols: bool,
         is_regex: bool,
     ) -> Result<String> {
-        let chat_client: Arc<dyn ChatClient> = build_chat_client_for(LlmUsage::ExplainCode, llm, self.container.data_dir())?;
+        let chat_client: Arc<dyn ChatClient> =
+            build_chat_client_for(LlmUsage::ExplainCode, llm, self.container.data_dir())?;
 
         let (token_tx, mut token_rx) = tokio::sync::mpsc::unbounded_channel::<String>();